==> rosmap_pipeline/6a_submodule_annotation.Rmd <==
---
title: "Cell Type Annotations for Submodules in ROSMAP"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

rm(list=ls())

library(dplyr)
library(ggplot2)
library(ggalluvial)
library(clusterProfiler)
library(org.Hs.eg.db)
library(RColorBrewer)
library(networkD3)

source("../utils/plot_theme.R")

COHORT <- "rosmap"

# Number of marker genes to use
NUM.MARKERS = 100
```

Cell type markers in the brain are taken from McKenzie et al. (Supplemental File 1, DOI:10.1038/s41598-018-27293-5). The study developed three different metrics to rank the relation of genes to specific cell types (enrichment, expression, specificity). 

```{r load_data}
# Get module genes for ROSMAP
modules.genes <- read.table("raw_data/Modules_March2018.txt", sep="\t", header=T, stringsAsFactors=F) %>% 
  subset(brainRegion == "DLPFC") %>%
  dplyr::select(Module=Module, Gene=GeneID, HGNC=external_gene_name)
modules <- unique(modules.genes$Module)
m.gene.sets <- lapply(modules, function(x) subset(modules.genes, Module==x)[,2:3])
names(m.gene.sets) <- modules

# Submodule genes for ROSMAP
# We will split these up by the module
submodules.genes <- readRDS("clean_data/3_iterative_WGCNA_cleaning/rosmap_submodule_sig_genes.RDS")

# Names of submodules
submodules <- unique(submodules.genes$Submodule)

# List of data frames by modules
gene.sets <- lapply(submodules, function(x) subset(submodules.genes, Submodule==x)[,2:3])
names(gene.sets) <- submodules

# Load data generated by McKenzie et al. (Supplemental File 1)
# https://doi.org/10.1038/s41598-018-27293-5
hum.enrich <- read.csv("raw_data/McKenzie_2017_Brain_Cell_Specific_Markers_Human_Enrich.csv", header=T, stringsAsFactors=F)
hum.express <- read.csv("raw_data/McKenzie_2017_Brain_Cell_Specific_Markers_Human_Expression.csv", header=T, stringsAsFactors=F)
hum.specific <- read.csv("raw_data/McKenzie_2017_Brain_Cell_Specific_Markers_Human_Specificity.csv", header=T, stringsAsFactors=F)

# Convert to lists based on cell type
# The authors recommend using the top 50 genes as markers
cell.types <- union(union(hum.enrich$Celltype, hum.express$Celltype), unique(hum.specific$Celltype))

hum.enrich <- lapply(cell.types, function(x) (subset(hum.enrich, Celltype == x) %>% dplyr::select(HGNC=gene, Grand.Mean=grand_mean))[1:NUM.MARKERS,])
hum.express <- lapply(cell.types, function(x) (subset(hum.express, Celltype == x) %>% dplyr::select(HGNC=gene, Grand.Rank=grand_rank))[1:NUM.MARKERS,])
hum.specific <- lapply(cell.types, function(x) (subset(hum.specific, Celltype == x) %>% dplyr::select(HGNC=gene, Grand.Mean=grand_mean))[1:NUM.MARKERS,])

cell.types <- c("Astroctyes", "Endothelial Cells", "Microglia", "Neurons", "Oligodendrocytes")

names(hum.enrich) <- cell.types
names(hum.express) <- cell.types
names(hum.specific) <- cell.types
```

Mapping HGNC to Ensembl
-----------------------

Cell type specific markers from McKenzie et al. are provided as HGNC symbols. I used `clusterProfiler` to map these to Ensembl accessions. 

```{r message=FALSE, warning=FALSE}
# McKenzie data set is using HGNC symbols
# Map to ENSEMBL symbols
hum.enrich <- lapply(cell.types, function(x) {
  gene.map <- bitr(hum.enrich[[x]]$HGNC, fromType="SYMBOL", toType="ENSEMBL", OrgDb=org.Hs.eg.db)
  hum.enrich[[x]] <- merge(hum.enrich[[x]], gene.map, by.x="HGNC", by.y="SYMBOL") %>% dplyr::arrange(desc(Grand.Mean))
})
names(hum.enrich) <- cell.types

hum.express <- lapply(cell.types, function(x) {
  gene.map <- bitr(hum.express[[x]]$HGNC, fromType="SYMBOL", toType="ENSEMBL", OrgDb=org.Hs.eg.db)
  hum.express[[x]] <- merge(hum.express[[x]], gene.map, by.x="HGNC", by.y="SYMBOL") %>% dplyr::arrange(desc(Grand.Rank))
})
names(hum.express) <- cell.types

hum.specific <- lapply(cell.types, function(x) {
  gene.map <- bitr(hum.specific[[x]]$HGNC, fromType="SYMBOL", toType="ENSEMBL", OrgDb=org.Hs.eg.db)
  hum.specific[[x]] <- merge(hum.specific[[x]], gene.map, by.x="HGNC", by.y="SYMBOL") %>% dplyr::arrange(desc(Grand.Mean))
})
names(hum.specific) <- cell.types
```

Cell Enrichment Score Ranks
---------------------------

I used the top markers from the cell enrichment score to annotate the submodules in the ROSMAP cohort. I performed a chi-square test to compare the frequency of cell type markers in any given submodule to the total frequency of markers used. I performed a similar analysis for the modules. 

The cell enrichment for a given gene is the log fold change when comparing the expression of the gene in the given cell type with the expression of the gene in all other cell types.

```{r warning=FALSE}
# Submodules
enrich.set <- lapply(submodules, function(s) {
  sapply(cell.types, function(x) hum.enrich[[x]]$ENSEMBL %in% gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(enrich.set) <- submodules
enrich.set <- do.call("rbind", enrich.set)

chi.sq.p <- sapply(submodules, function(s) chisq.test(rbind(enrich.set[s,], sapply(hum.enrich, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data <- as.data.frame(stack(enrich.set))[,c(1,2,4)]
colnames(plot.data) <- c("Submodule", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Submodule, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=submodules, y=rowSums(enrich.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Submodule") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/hum.enrich.annot.svg", plot=p, width=8, height=7.5)

# Modules
enrich.set <- lapply(modules, function(s) {
  sapply(cell.types, function(x) hum.enrich[[x]]$ENSEMBL %in% m.gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(enrich.set) <- modules
enrich.set <- do.call("rbind", enrich.set)

chi.sq.p <- sapply(modules, function(s) chisq.test(rbind(enrich.set[s,], sapply(hum.enrich, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data <- as.data.frame(stack(enrich.set))[,c(1,2,4)]
colnames(plot.data) <- c("Module", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Module, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=modules, y=rowSums(enrich.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Module") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/m.hum.enrich.annot.svg", plot=p, width=5, height=7.5)
```

Cell Expression Score Ranks
---------------------------

I used the top markers from the cell expression score to annotate the submodules in the ROSMAP cohort. I performed a chi-square test to compare the frequency of cell type markers in any given submodule to the total frequency of markers used. I performed a similar analysis for the modules. 

The cell expression for a given gene is the absolute expression value for the given gene in a given cell type.

```{r warning=FALSE}
# Submodules
express.set <- lapply(submodules, function(s) {
  sapply(cell.types, function(x) hum.express[[x]]$ENSEMBL %in% gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(express.set) <- submodules
express.set <- do.call("rbind", express.set)

chi.sq.p <- sapply(submodules, function(s) chisq.test(rbind(express.set[s,], sapply(hum.express, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data <- as.data.frame(stack(express.set))[,c(1,2,4)]
colnames(plot.data) <- c("Submodule", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Submodule, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=submodules, y=rowSums(express.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Submodule") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/hum.express.annot.svg", plot=p, width=8, height=7.5)

# Modules
express.set <- lapply(modules, function(s) {
  sapply(cell.types, function(x) hum.express[[x]]$ENSEMBL %in% m.gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(express.set) <- modules
express.set <- do.call("rbind", express.set)

chi.sq.p <- sapply(modules, function(s) chisq.test(rbind(express.set[s,], sapply(hum.express, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data <- as.data.frame(stack(express.set))[,c(1,2,4)]
colnames(plot.data) <- c("Module", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Module, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=modules, y=rowSums(express.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Module") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/m.hum.express.annot.svg", plot=p, width=5, height=7.5)
```

Cell Specificity Score Ranks
---------------------------

I used the top markers from the cell specificity score to annotate the submodules in the ROSMAP cohort. I performed a chi-square test to compare the frequency of cell type markers in any given submodule to the total frequency of markers used. I performed a similar analysis for the modules. 

The cell specificity for a given gene is the minimum log fold change when pairwise comparing the expression of the gene in the given cell type with the expression of the gene in all other cell types.

```{r warning=FALSE}
# Submodules
specific.set <- lapply(submodules, function(s) {
  sapply(cell.types, function(x) hum.specific[[x]]$ENSEMBL %in% gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(specific.set) <- submodules
specific.set <- do.call("rbind", specific.set)

chi.sq.p <- sapply(submodules, function(s) chisq.test(rbind(specific.set[s,], sapply(hum.specific, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data.1 <- as.data.frame(stack(specific.set))[,c(1,2,4)]
colnames(plot.data.1) <- c("Submodule", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data.1, aes(x=Submodule, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=submodules, y=rowSums(specific.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Submodule") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/hum.specific.annot.svg", plot=p, width=8, height=7.5)

# Modules
specific.set <- lapply(modules, function(s) {
  sapply(cell.types, function(x) hum.specific[[x]]$ENSEMBL %in% m.gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(specific.set) <- modules
specific.set <- do.call("rbind", specific.set)

chi.sq.p <- sapply(modules, function(s) chisq.test(rbind(specific.set[s,], sapply(hum.specific, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data.2 <- as.data.frame(stack(specific.set))[,c(1,2,4)]
colnames(plot.data.2) <- c("Module", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data.2, aes(x=Module, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=modules, y=rowSums(specific.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Module") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/m.hum.specific.annot.svg", plot=p, width=5, height=7.5)
```

Combined Graph for Specificity
------------------------------

I found that the cell specificity markers were the best for demonstrating the cell type specificity of the submodules. I created a graphic that combines the modules and submodules into one image.

```{r}
plot.data <- rbind(
  plot.data.2 %>%
    dplyr::mutate(Scope="Module"),
  plot.data.1 %>% 
    dplyr::select(Module=Submodule, everything()) %>%
    dplyr::mutate(Scope="Submodule")
)
plot.data$Module <- factor(plot.data$Module, levels=c(modules, submodules))

p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Module, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(
    axis.text.x=element_text(angle=60, hjust=1),
    legend.position="bottom"
  )
print(p)
ggsave(filename="results/6_submodule_annotation/cellspecific.annot.svg", plot=p, width=13, height=8)
```

Flow Diagrams
-------------

I created a flow diagram to make it clear how genes from modules become much more cell-specific when grouped using the `iterativeWGCNA` submodules.

```{r}
# Tables of gene by cell type
sm.cell.genes <- lapply(submodules, function(s) {
  do.call(rbind, lapply(cell.types, function(x) {
    tbl <- as.data.frame(list(hum.specific[[x]]$ENSEMBL[hum.specific[[x]]$ENSEMBL %in% gene.sets[[s]]$Gene])) %>%
      dplyr::mutate(Cell.Type=x)
    colnames(tbl) <- c("Gene", "Cell.Type")
    return(tbl)
  }))
})
names(sm.cell.genes) <- submodules

# Table of gene by module by cell type
m.cell.genes <- lapply(modules, function(s) {
  do.call(rbind, lapply(cell.types, function(x) {
    tbl <- as.data.frame(list(hum.specific[[x]]$ENSEMBL[hum.specific[[x]]$ENSEMBL %in% m.gene.sets[[s]]$Gene])) %>%
      dplyr::mutate(Cell.Type=x)
    colnames(tbl) <- c("Gene", "Cell.Type")
    return(tbl)
  }))
})
names(m.cell.genes) <- modules

plot.data <- as.data.frame(do.call(
  rbind,
  lapply(modules, function(module) {
    do.call(
      rbind,
      lapply(submodules, function(submodule) {
        do.call(
          rbind,
          lapply(cell.types, function(cell.type) {
            m.g <- m.cell.genes[[module]] %>% dplyr::filter(Cell.Type==cell.type)
            sm.g <- sm.cell.genes[[submodule]] %>% dplyr::filter(Cell.Type==cell.type)
            return(c(module, submodule, cell.type, length(intersect(m.g$Gene, sm.g$Gene))))
          })
        )
      })
    )
  })
))
colnames(plot.data) <- c("Module", "Submodule", "Cell.Type", "Freq")
plot.data$Freq <- as.numeric(as.character(plot.data$Freq))

is_alluvia_form(plot.data, axes=1:3, silent=TRUE)

ggplot(plot.data, aes(y=Freq, axis1=Module, axis2=Submodule)) +
  geom_alluvium(aes(fill=Cell.Type), width=1/12) +
  geom_stratum(width=1/12, fill="white", color="black") +
  geom_label(stat="stratum", label.strata=TRUE) +
  scale_x_discrete(limits=c("Module", "Submodule"), expand=c(0.05, 0.05)) +
  scale_fill_brewer(type="qual", palette="Set1") + 
  coord_flip() +
  basic_theme
ggsave(filename="results/6_submodule_annotation/cell.specific.alluvial.svg", plot=p, width=13, height=8)
```

The Alluvial diagram is not really helpful because the strata are too close together and it is not obvious what is going on intuitively. I tried using Sankey diagrams instead, which are similar but separate the strata out.

```{r}
plot.data <- plot.data[plot.data$Freq != 0,]
plot.data$Cell.Type <- substring(as.character(plot.data$Cell.Type), 1, 1)

nodes <- data.frame(Name=unique(c(as.character(plot.data$Module), as.character(plot.data$Submodule))))
nodes$Group = sapply(strsplit(as.character(nodes$Name), "_"), function(x) x[1])

plot.data$Module <- sapply(as.character(plot.data$Module), function(x) which(as.character(nodes$Name) == x) - 1)
plot.data$Submodule <- sapply(as.character(plot.data$Submodule), function(x) which(as.character(nodes$Name) == x) - 1)

node.colors <- paste0(
  "d3.scaleOrdinal() .domain([",
  paste0("'", c("A", "E", "M", "N", "O"), "'", collapse=","),
  "]) .range([ ",
  paste0("'", brewer.pal(5, "Set3"), "'", collapse=","), "])"
)

net <- sankeyNetwork(
  Links=plot.data, Nodes=nodes, Source="Module", Target="Submodule",
  Value="Freq", NodeID="Name", fontSize=20, fontFamily="Arial", nodeWidth=50,
  LinkGroup="Cell.Type", NodeGroup="Group", colourScale=node.colors
)
saveNetwork(net, file="~/Desktop/AMPAD_Submodules/rosmap_pipeline/results/6_submodule_annotation/cell.specific.sankey.html", selfcontained=TRUE)
```
